sync

+++ test/smoke/src/main.ts
@@ -4,29 +4,79 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as fs from 'fs';
-import * as https from 'https';
 import * as cp from 'child_process';
 import * as path from 'path';
 import * as minimist from 'minimist';
 import * as tmp from 'tmp';
 import * as rimraf from 'rimraf';
 import * as mkdirp from 'mkdirp';
+import { ncp } from 'ncp';
+import {
+	Application,
+	Quality,
+	ApplicationOptions,
+	MultiLogger,
+	Logger,
+	ConsoleLogger,
+	FileLogger,
+} from '../../automation';
 
-const tmpDir = tmp.dirSync() as { name: string; removeCallback: Function; };
+import { setup as setupDataMigrationTests } from './areas/workbench/data-migration.test';
+import { setup as setupDataLossTests } from './areas/workbench/data-loss.test';
+import { setup as setupDataExplorerTests } from './areas/explorer/explorer.test';
+import { setup as setupDataPreferencesTests } from './areas/preferences/preferences.test';
+import { setup as setupDataSearchTests } from './areas/search/search.test';
+import { setup as setupDataCSSTests } from './areas/css/css.test';
+import { setup as setupDataEditorTests } from './areas/editor/editor.test';
+import { setup as setupDataDebugTests } from './areas/debug/debug.test';
+import { setup as setupDataGitTests } from './areas/git/git.test';
+import { setup as setupDataStatusbarTests } from './areas/statusbar/statusbar.test';
+import { setup as setupDataExtensionTests } from './areas/extensions/extensions.test';
+import { setup as setupTerminalTests } from './areas/terminal/terminal.test';
+import { setup as setupDataMultirootTests } from './areas/multiroot/multiroot.test';
+import { setup as setupDataLocalizationTests } from './areas/workbench/localization.test';
+import { setup as setupLaunchTests } from './areas/workbench/launch.test';
+
+if (!/^v10/.test(process.version)) {
+	console.error('Error: Smoketest must be run using Node 10. Currently running', process.version);
+	process.exit(1);
+}
+
+const tmpDir = tmp.dirSync({ prefix: 't' }) as { name: string; removeCallback: Function; };
 const testDataPath = tmpDir.name;
 process.once('exit', () => rimraf.sync(testDataPath));
 
 const [, , ...args] = process.argv;
-const opts = minimist(args, { string: ['build', 'stable-build', 'screenshots'] });
+const opts = minimist(args, {
+	string: [
+		'build',
+		'stable-build',
+		'wait-time',
+		'test-repo',
+		'screenshots',
+		'log'
+	],
+	boolean: [
+		'verbose',
+		'remote',
+		'web',
+		'headless'
+	],
+	default: {
+		verbose: false
+	}
+});
 
-opts.screenshots = opts.screenshots === '' ? path.join(testDataPath, 'screenshots') : opts.screenshots;
-
-const workspacePath = path.join(testDataPath, 'smoketest.code-workspace');
 const testRepoUrl = 'https://github.com/Microsoft/vscode-smoketest-express';
-const testRepoLocalDir = path.join(testDataPath, 'vscode-smoketest-express');
-const keybindingsPath = path.join(testDataPath, 'keybindings.json');
+const workspacePath = path.join(testDataPath, 'vscode-smoketest-express');
 const extensionsPath = path.join(testDataPath, 'extensions-dir');
 mkdirp.sync(extensionsPath);
+
+const screenshotsPath = opts.screenshots ? path.resolve(opts.screenshots) : null;
+
+if (screenshotsPath) {
+	mkdirp.sync(screenshotsPath);
+}
 
 function fail(errorMessage): void {
 	console.error(errorMessage);
@@ -55,142 +105,197 @@ function getDevElectronPath(): string {
 	}
 }
 
+function getBuildElectronPath(root: string): string {
+	switch (process.platform) {
+		case 'darwin':
+			return path.join(root, 'Contents', 'MacOS', 'Electron');
+		case 'linux': {
+			const product = require(path.join(root, 'resources', 'app', 'product.json'));
+			return path.join(root, product.applicationName);
+		}
+		case 'win32': {
+			const product = require(path.join(root, 'resources', 'app', 'product.json'));
+			return path.join(root, `${product.nameShort}.exe`);
+		}
+		default:
+			throw new Error('Unsupported platform.');
+	}
+}
+
 let testCodePath = opts.build;
 let stableCodePath = opts['stable-build'];
+let electronPath: string;
+let stablePath: string | undefined = undefined;
 
 if (testCodePath) {
-	process.env.VSCODE_PATH = testCodePath;
+	electronPath = getBuildElectronPath(testCodePath);
 
 	if (stableCodePath) {
-		process.env.VSCODE_STABLE_PATH = stableCodePath;
+		stablePath = getBuildElectronPath(stableCodePath);
 	}
 } else {
 	testCodePath = getDevElectronPath();
-	process.env.VSCODE_PATH = testCodePath;
+	electronPath = testCodePath;
 	process.env.VSCODE_REPOSITORY = repoPath;
 	process.env.VSCODE_DEV = '1';
 	process.env.VSCODE_CLI = '1';
 }
 
-if (!fs.existsSync(testCodePath)) {
-	fail(`Can't find Code at ${testCodePath}.`);
+if (!opts.web && !fs.existsSync(electronPath || '')) {
+	fail(`Can't find Code at ${electronPath}.`);
 }
 
-process.env.VSCODE_USER_DIR = path.join(testDataPath, 'user-dir');
-process.env.VSCODE_EXTENSIONS_DIR = extensionsPath;
-process.env.SMOKETEST_REPO = testRepoLocalDir;
-process.env.VSCODE_WORKSPACE_PATH = workspacePath;
-process.env.VSCODE_KEYBINDINGS_PATH = keybindingsPath;
-process.env.SCREENSHOTS_DIR = opts.screenshots || '';
+if (typeof stablePath === 'string' && !fs.existsSync(stablePath)) {
+	fail(`Can't find Stable Code at ${stablePath}.`);
+}
 
+const userDataDir = path.join(testDataPath, 'd');
+
+let quality: Quality;
 if (process.env.VSCODE_DEV === '1') {
-	process.env.VSCODE_EDITION = 'dev';
-} else if ((testCodePath.indexOf('Code - Insiders') /* macOS/Windows */ || testCodePath.indexOf('code-insiders') /* Linux */) >= 0) {
-	process.env.VSCODE_EDITION = 'insiders';
+	quality = Quality.Dev;
+} else if (electronPath.indexOf('Code - Insiders') >= 0 /* macOS/Windows */ || electronPath.indexOf('code-insiders') /* Linux */ >= 0) {
+	quality = Quality.Insiders;
 } else {
-	process.env.VSCODE_EDITION = 'stable';
+	quality = Quality.Stable;
 }
 
-function getKeybindingPlatform(): string {
-	switch (process.platform) {
-		case 'darwin': return 'osx';
-		case 'win32': return 'win';
-		default: return process.platform;
-	}
-}
+async function setupRepository(): Promise<void> {
+	if (opts['test-repo']) {
+		console.log('*** Copying test project repository:', opts['test-repo']);
+		rimraf.sync(workspacePath);
+		// not platform friendly
+		if (process.platform === 'win32') {
+			cp.execSync(`xcopy /E "${opts['test-repo']}" "${workspacePath}"\\*`);
+		} else {
+			cp.execSync(`cp -R "${opts['test-repo']}" "${workspacePath}"`);
+		}
 
-function toUri(path: string): string {
-	if (process.platform === 'win32') {
-		return `file:///${path.replace(/\\/g, '/')}`;
-	}
+	} else {
+		if (!fs.existsSync(workspacePath)) {
+			console.log('*** Cloning test project repository...');
+			cp.spawnSync('git', ['clone', testRepoUrl, workspacePath]);
+		} else {
+			console.log('*** Cleaning test project repository...');
+			cp.spawnSync('git', ['fetch'], { cwd: workspacePath });
+			cp.spawnSync('git', ['reset', '--hard', 'FETCH_HEAD'], { cwd: workspacePath });
+			cp.spawnSync('git', ['clean', '-xdf'], { cwd: workspacePath });
+		}
 
-	return `file://${path}`;
+		console.log('*** Running yarn...');
+		cp.execSync('yarn', { cwd: workspacePath, stdio: 'inherit' });
+	}
 }
 
 async function setup(): Promise<void> {
 	console.log('*** Test data:', testDataPath);
 	console.log('*** Preparing smoketest setup...');
 
-	const keybindingsUrl = `https://raw.githubusercontent.com/Microsoft/vscode-docs/master/scripts/keybindings/doc.keybindings.${getKeybindingPlatform()}.json`;
-	console.log('*** Fetching keybindings...');
-
-	await new Promise((c, e) => {
-		https.get(keybindingsUrl, res => {
-			const output = fs.createWriteStream(keybindingsPath);
-			res.on('error', e);
-			output.on('error', e);
-			output.on('close', c);
-			res.pipe(output);
-		}).on('error', e);
-	});
-
-	if (!fs.existsSync(workspacePath)) {
-		console.log('*** Creating workspace file...');
-		const workspace = {
-			id: (Date.now() + Math.round(Math.random() * 1000)).toString(),
-			folders: [
-				toUri(path.join(testRepoLocalDir, 'public')),
-				toUri(path.join(testRepoLocalDir, 'routes')),
-				toUri(path.join(testRepoLocalDir, 'views'))
-			]
-		};
-
-		fs.writeFileSync(workspacePath, JSON.stringify(workspace, null, '\t'));
-	}
-
-	if (!fs.existsSync(testRepoLocalDir)) {
-		console.log('*** Cloning test project repository...');
-		cp.spawnSync('git', ['clone', testRepoUrl, testRepoLocalDir]);
-	} else {
-		console.log('*** Cleaning test project repository...');
-		cp.spawnSync('git', ['fetch'], { cwd: testRepoLocalDir });
-		cp.spawnSync('git', ['reset', '--hard', 'FETCH_HEAD'], { cwd: testRepoLocalDir });
-		cp.spawnSync('git', ['clean', '-xdf'], { cwd: testRepoLocalDir });
-	}
-
-	console.log('*** Running npm install...');
-	cp.execSync('npm install', { cwd: testRepoLocalDir, stdio: 'inherit' });
+	await setupRepository();
 
 	console.log('*** Smoketest setup done!\n');
 }
 
-/**
- * WebDriverIO 4.8.0 outputs all kinds of "deprecation" warnings
- * for common commands like `keys` and `moveToObject`.
- * According to https://github.com/Codeception/CodeceptJS/issues/531,
- * these deprecation warnings are for Firefox, and have no alternative replacements.
- * Since we can't downgrade WDIO as suggested (it's Spectron's dep, not ours),
- * we must suppress the warning with a classic monkey-patch.
- *
- * @see webdriverio/lib/helpers/depcrecationWarning.js
- * @see https://github.com/webdriverio/webdriverio/issues/2076
- */
-// Filter out the following messages:
-const wdioDeprecationWarning = /^WARNING: the "\w+" command will be depcrecated soon./; // [sic]
-// Monkey patch:
-const warn = console.warn;
-console.warn = function suppressWebdriverWarnings(message) {
-	if (wdioDeprecationWarning.test(message)) { return; }
-	warn.apply(console, arguments);
-};
+function createOptions(): ApplicationOptions {
+	const loggers: Logger[] = [];
+
+	if (opts.verbose) {
+		loggers.push(new ConsoleLogger());
+	}
+
+	let log: string | undefined = undefined;
+
+	if (opts.log) {
+		loggers.push(new FileLogger(opts.log));
+		log = 'trace';
+	}
+	return {
+		quality,
+		codePath: opts.build,
+		workspacePath,
+		userDataDir,
+		extensionsPath,
+		waitTime: parseInt(opts['wait-time'] || '0') || 20,
+		logger: new MultiLogger(loggers),
+		verbose: opts.verbose,
+		log,
+		screenshotsPath,
+		remote: opts.remote,
+		web: opts.web,
+		headless: opts.headless
+	};
+}
 
 before(async function () {
 	// allow two minutes for setup
 	this.timeout(2 * 60 * 1000);
 	await setup();
+	this.defaultOptions = createOptions();
 });
 
-// import './areas/workbench/data-migration.test';
-import './areas/workbench/data-loss.test';
-import './areas/explorer/explorer.test';
-import './areas/preferences/preferences.test';
-import './areas/search/search.test';
-import './areas/multiroot/multiroot.test';
-import './areas/css/css.test';
-import './areas/editor/editor.test';
-import './areas/debug/debug.test';
-import './areas/git/git.test';
-import './areas/terminal/terminal.test';
-import './areas/statusbar/statusbar.test';
-import './areas/extensions/extensions.test';
-import './areas/workbench/localization.test';
+after(async function () {
+	await new Promise(c => setTimeout(c, 500)); // wait for shutdown
+
+	if (opts.log) {
+		const logsDir = path.join(userDataDir, 'logs');
+		const destLogsDir = path.join(path.dirname(opts.log), 'logs');
+		await new Promise((c, e) => ncp(logsDir, destLogsDir, err => err ? e(err) : c()));
+	}
+
+	await new Promise((c, e) => rimraf(testDataPath, { maxBusyTries: 10 }, err => err ? e(err) : c()));
+});
+
+if (!opts.web) {
+	setupDataMigrationTests(stableCodePath, testDataPath);
+}
+
+describe('Running Code', () => {
+	before(async function () {
+		const app = new Application(this.defaultOptions);
+		await app!.start(opts.web ? false : undefined);
+		this.app = app;
+	});
+
+	after(async function () {
+		await this.app.stop();
+	});
+
+	if (screenshotsPath) {
+		afterEach(async function () {
+			if (this.currentTest.state !== 'failed') {
+				return;
+			}
+			const app = this.app as Application;
+			const name = this.currentTest.fullTitle().replace(/[^a-z0-9\-]/ig, '_');
+
+			await app.captureScreenshot(name);
+		});
+	}
+
+	if (opts.log) {
+		beforeEach(async function () {
+			const app = this.app as Application;
+			const title = this.currentTest.fullTitle();
+
+			app.logger.log('*** Test start:', title);
+		});
+	}
+
+	if (!opts.web) { setupDataLossTests(); }
+	setupDataExplorerTests();
+	if (!opts.web) { setupDataPreferencesTests(); }
+	setupDataSearchTests();
+	setupDataCSSTests();
+	setupDataEditorTests();
+	if (!opts.web) { setupDataDebugTests(); }
+	setupDataGitTests();
+	setupDataStatusbarTests(!!opts.web);
+	setupDataExtensionTests();
+	setupTerminalTests();
+	if (!opts.web) { setupDataMultirootTests(); }
+	setupDataLocalizationTests();
+});
+
+if (!opts.web) {
+	setupLaunchTests();
+}
